Reorganizando el codigo
