--- telegram.token
@@ -1,1 +1,0 @@
-143179136:AAHQKOCWGAbPvlL5loKqI2lyyVopktargM0
